Add an alert that tells a user a card was added
When a user adds a new card, the term is stored in a session.

That session value is then rendered in a partial when the user is
redirected to make a new card.

The session[:last_card_term] value becomes nil when the user navigates
away from the page. This way, the user cannot add a card, do something else,
go to add a new card to a different deck, and still see the term that was last added.