Tests no longer working
When we added complexity to our program--adding users as authors to the
deck, the tests failed. I have not been able to figure out how to make them
work.

I read about Factory Girl and Database Cleaner and tried getting them to
work for a couple of hours to no avail. I cannot generate unique names, so
our model validations make all of the tests fail. When the validation is removed,
it says that nil is not a symbol. I cannot solve this at the current time, and
I may come back to it later.